requirements.txt
# Use specific versions for reproducibility
# Check for the latest compatible versions if needed
discord.py==2.3.2
openai==1.30.1
# Optional, but useful for loading .env files during local development
python-dotenv==1.0.1
python-dotenv>=0.21.0
